examples/pytorch/sdxl/requirements_sdxl.txt
accelerate>=0.22.0
datasets
ftfy
Jinja2
tensorboard
torchvision
transformers>=4.25.1
